feat: Added CRUD for schools, documented the school route, created the roles route, added roles to user responses, improved school and user responses, added the role route, and finalized school documentation on Swagger.

+++ apps/school/services/school.py
@@ -25,11 +25,14 @@ def school_update(*,school: School, data: dict) -> School:
     fields = ['name', 'address','responsible',"geolocation"]
     school, has_updated = model_update(instance=school, fields=fields, data=data)
     
-    return school
+    return school , has_updated
 
 @transaction.atomic
 def school_delete(*,id:str) -> School:
     school = school_get(school=School, id=id)
+    if school is None:
+        return school
+    
     school.full_clean()
     school.delete()
